Use gomock to test that validator calls the trafficclaim database
Problem: We're not unit testing that the validator calls the
trafficclaim database as we expect.

Solution: Add unit tests that verify we correctly interpret "host
claims", "port claims" and so on.  Add unit tests that send a "golden"
VirtualService spec at the validator to confirm that the
half-Kubernetes, half-Istio unmarshaling works as expected.