Add TrafficClaims resolution
Problem: VirtualServices are accepted or rejected based on drawing a
random number because there's no facility to examine TrafficClaims and
see if they grant ability to configure certain hosts.

Solution: Add trafficclaim.go, which provides a claim Db that can
examine TrafficClaims in Kubernetes to see if particular
hosts/ports/paths should be allowed or not.