Only get TrafficClaims once per validation
Problem: Before we were getting the TrafficClaims in a namespace for
each host/port/path config tuple.

Solution: Any TrafficClaim in a namespace might grant the config tuple
in question, so we must check each tuple against all of the
TrafficClaims.  We also must only check the TrafficClaims in this
namespace, so we don't need a global view of all TrafficClaims anywhere.

Create a new Verification interface - a Verification is one run against
the database.  Currently, a Verification calls the kubernetes client
List() to get the TrafficClaims in a namespace - someday it could
instead just fetch the list out of a cache if that turns out to be a
better pattern.  A Verification is just a container for a set of
TrafficClaims so that we can then evaluate config tuples over and over
against that set.

Update unit tests to use the Verification.